refactor(#4): extract AddressSearchInfo to own component
Reason: for reuse in /five-days
